# docs/swagger/facebook-ads-swagger.yaml
openapi: 3.0.3
info:
  title: Marketing Hub - Facebook Ads API
  version: 1.0.0
  description: |
    Contrato dos endpoints do backend que atendem a integração com Facebook Ads.
    A implementação canônica dos controllers fica concentrada no pacote
    `com.marketinghub.facebookads.controller`, mantendo as rotas públicas legadas
    para compatibilidade com frontend e workers.
...
paths:
...
  /api/accounts/facebook/{id}:
    put:
      tags: [Facebook Accounts]
      summary: Atualiza uma conta do Facebook sem descartar campos não enviados pela UI.
      parameters: [ { name: id, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Conta atualizada. } }
    delete:
      tags: [Facebook Accounts]
      summary: Remove uma conta do Facebook.
      parameters: [ { name: id, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '204': { description: Conta removida. } }
...
  /api/accounts/facebook/{accountId}/pages:
    get:
      tags: [Facebook Pages]
      summary: Lista páginas vinculadas à conta Facebook.
      parameters: [ { name: accountId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Páginas retornadas. } }
    post:
      tags: [Facebook Pages]
      summary: Vincula uma página à conta Facebook.
      parameters: [ { name: accountId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Página criada. } }
  /api/accounts/facebook/{accountId}/pages/{pageRecordId}:
    put:
      tags: [Facebook Pages]
      summary: Atualiza uma página vinculada à conta Facebook.
      parameters:
        - { name: accountId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: pageRecordId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses: { '200': { description: Página atualizada. } }
    delete:
      tags: [Facebook Pages]
      summary: Remove uma página vinculada quando não houver experimentos dependentes.
      parameters:
        - { name: accountId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: pageRecordId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses: { '204': { description: Página removida. } }
...